extract muiTheme and formikConfig to its own files
