+++ src/muiMainTheme.ts
@@ -1,0 +1,11 @@
+import { createMuiTheme } from '@material-ui/core/styles';
+
+const theme = createMuiTheme({
+    palette: {
+        primary: {
+            main: '#f4bf41'
+        }
+    }
+});
+
+export default theme;
